requirements.txt
aiofiles==24.1.0
alembic==1.16.4
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.1
async-timeout==5.0.1
asyncpg==0.30.0
billiard==4.2.1
celery==5.5.3
certifi==2025.7.14
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
exceptiongroup==1.3.0
fastapi==0.116.1
flower==2.0.1
gevent==25.5.1
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
humanize==4.12.3
idna==3.10
kombu==5.5.4
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
pillow==11.3.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
redis==5.2.1
replicate==1.0.7
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlmodel==0.0.24
starlette==0.47.2
tomli==2.2.1
tornado==6.5.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uvicorn==0.35.0
vine==5.1.0
wcwidth==0.2.13
zope.event==5.1.1
zope.interface==7.2
